fix interface to match reason-react
This modifies the interfaces of the React module to match the one
exposed by reason-react, based on a few issues I ran into so far. The
use of Stdlib.ref is replaced with a one-off record that has a `current`
field instead of `contents`, `cloneElement` loses its children
parameter, since it is only meant to modify the attributes, and the
provider method of `Context` uses its internal `t` type.